backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.10.0
pinecone-client==3.0.0
python-multipart==0.0.6
pypdf==3.17.4
tiktoken==0.5.2
httpx==0.27.0